WIP refactor how weapons work
TLDR of this is that models and views need to be specialized based upon whether a weapon is a melee, missile, or magic. So the single Weapon model becomes three and the single Weapon view becomes three as well.

This isn't perfect code yet but it's the scaffolding we need.